+++ extensions/1bookmark/src/views/LoginView.tsx
@@ -1,35 +1,39 @@
 import { useEffect, useRef, useState } from "react";
-import { ActionPanel, Action, Form } from "@raycast/api";
+import { ActionPanel, Action, Form, Icon } from "@raycast/api";
 import { CachedQueryClientProvider } from "../components/CachedQueryClientProvider";
 import { trpc } from "@/utils/trpc.util";
 import { handleSignIn } from "@/handle-signin";
 import { useAtom } from "jotai";
 import { sessionTokenAtom } from "@/states/session-token.state";
+import { loggingEmailAtom, loggingTokenSentAtom } from "../states/logging-email.state";
 
 function Body() {
   const [, setSessionToken] = useAtom(sessionTokenAtom);
-  const [sentToken, setSentToken] = useState(false);
   const { mutateAsync, isPending } = trpc.login.generateMagicLink.useMutation();
   const verificationTokenRef = useRef<Form.TextField>(null);
 
-  const [email, setEmail] = useState("");
+  const [tokenSent, setTokenSent] = useAtom(loggingTokenSentAtom);
+  const [email, setEmail] = useAtom(loggingEmailAtom);
+
   const [code, setCode] = useState("");
+
+  const emailRef = useRef<Form.TextField>(null);
 
   const [isLoginPending, setIsLoginPending] = useState(false);
   const isLoading = isPending || isLoginPending;
   const requestToToken = async (email: string) => {
     await mutateAsync({ email });
 
-    setSentToken(true);
+    setTokenSent(true);
   };
 
   useEffect(() => {
-    if (!sentToken) {
+    if (!tokenSent) {
       return;
     }
 
     verificationTokenRef.current?.focus();
-  }, [sentToken]);
+  }, [tokenSent]);
 
   return (
     <Form
@@ -37,20 +41,23 @@ function Body() {
       actions={
         <ActionPanel>
           <Action.SubmitForm
-            title={sentToken && email ? "Login" : "Send Login Code to Email"}
+            icon={Icon.Person}
+            title={tokenSent && email ? "Login" : "Send Login Code to Email"}
             onSubmit={async () => {
-              if (!sentToken && email) {
+              if (!tokenSent && email) {
                 requestToToken(email);
                 return;
               }
 
-              if (sentToken && email && code) {
+              if (tokenSent && email && code) {
                 setIsLoginPending(true);
                 await handleSignIn({
                   email,
                   token: code,
                   onSuccess: (sessionToken: string) => {
                     setSessionToken(sessionToken);
+                    setTokenSent(false);
+                    setEmail("");
                     // Onboarding 뷰를 거치지 않고 사용할 땐 pop하지 않는다.
                     // pop()
                   },
@@ -59,18 +66,40 @@ function Body() {
               }
             }}
           />
+          {tokenSent && (
+            <Action
+              title="Reset Email"
+              icon={Icon.ArrowLeft}
+              shortcut={{ modifiers: ["cmd"], key: "r" }}
+              onAction={() => {
+                setTokenSent(false);
+                setEmail("");
+
+                setTimeout(() => {
+                  emailRef.current?.focus();
+                }, 100);
+              }}
+            />
+          )}
         </ActionPanel>
       }
     >
-      {!sentToken && (
+      {!tokenSent && (
         <>
           <Form.Description text="👋🏼 Input Email to Login" />
-          <Form.TextField id="email" title="Email" placeholder="Email" onChange={(e) => setEmail(e)} />
+          <Form.TextField
+            ref={emailRef}
+            id="email"
+            title="Email"
+            placeholder="Email"
+            onChange={(e) => setEmail(e)}
+            autoFocus={true}
+          />
           <Form.Description text='Press "Command(⌘) + Enter"' />
         </>
       )}
 
-      {sentToken && (
+      {tokenSent && (
         <>
           <Form.Description text={`Login code sent to ${email} email.`} />
           <Form.Description text={`Enter the 6-digit login code sent to your email.`} />
@@ -81,7 +110,8 @@ function Body() {
             placeholder="Verification Token"
             onChange={(e) => setCode(e)}
           />
-          <Form.Description text='Press "Command(⌘) + Enter" again to login' />
+          <Form.Description text='Press "Command(⌘) + Enter" again to login.' />
+          <Form.Description text='Or press "Command(⌘) + R" to reset email.' />
         </>
       )}
     </Form>